Clean up projects tab management
